face_recognizer_portfolio_milestone/requirements.txt
face_recognition==1.2.2
pillow